changed ps scripts

+++ Crabot/stop-application.ps1
@@ -1,5 +1,9 @@
 $Crabot_process = Get-Process -Name Crabot
 
-if ($Crabot_process.Id -ne $null) {
+if ($Crabot_process -ne $null && $Crabot_process.Id -ne $null) {
+	Write-Output "Stopping process 'Crabot' with PID"
+	Write-Output $Crabot_process.Id
 	Stop-Process $Crabot_process.Id
 }
+
+Read-Host
